+++ packages/server/customer-os-common-module/services/init.go
@@ -312,7 +312,7 @@ func InitCommonServices(
 	agentImpl.SetListeners(agentListenerImpl)
 
 	// initialize registration service
-	registrationImpl := registration.NewRegistrationService(eventsImpl, postgresRepositories, neo4jRepositories, contactImpl, emailImpl, flowImpl, mailboxImpl, orgImpl, postmarkImpl, userImpl, agentImpl)
+	registrationImpl := registration.NewRegistrationService(eventsImpl, postgresRepositories, neo4jRepositories, contactImpl, emailImpl, flowImpl, mailboxImpl, mailstackImpl, orgImpl, postmarkImpl, userImpl, agentImpl)
 
 	// Initialize CommonServices struct
 	common := CommonServices{
